Add web display of output.
It is difficult for users to view the cost estimation of a large number
of resources in the command line, plain text or JSON format. To make
this process easier, a web output is generated as a HTML file with
multiple collapsing tables which are considerably easier to identify,
scroll through and read.