Added correct tier rate determination.
Some SKUs have multiple tier rates and initially only the first was
chosen. Now, tier rate is determined via a parsed function that
determines whether the tiered rate is the good one. For core and
RAM instances, the function always returns true (there is only one tier
rate). For disks, the function returns true for a startUsageAmount <=
size of the disk. This verification is done from the tier rate with the
highest startUsageAmount to the lowest. Aditionally cleaned
ResourceState interface and errors from ComputeDiskState.GetWebTables().